ImmediateSmallerElement.cpp
This is the solution to the problem in which :
Given an integer array of size N. For each element in the array, check whether the right adjacent element (on the next immediate position) of the array is smaller. If next element is smaller, print that element. If not, then print -1.